Commented out an incomplete block of code
Commented out a check to cover the case of installing 32bit java
on 64 bit Windows: the path for java_home would change.

I did'nt know exactly how to check for that from node data and I
did'nt needed this anyway, but I've left the code to show where/how
that could be done.